Moved some file for better organization.
