+++ src/com/archosResearch/jCHEKS/communicator/tcp/TCPCommunicator.java
@@ -5,6 +5,7 @@
  */
 package com.archosResearch.jCHEKS.communicator.tcp;
 
+import com.archosResearch.jCHEKS.communicator.AbstractCommunicator;
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
