--- src/com/archosResearch/jCHEKS/communicator/tcp/Temp.java
@@ -1,11 +1,0 @@
-package com.archosResearch.jCHEKS.communicator.tcp;
-
-
-/**
- *
- * @author fgagnon
- */
-public class Temp {
-
-}
-
